Add lots of functionality
This commit adds tests with VacheTache.

Thic commit adds Now and UtcNow and the possibility to set the same.
I am not happy with having to Set... methods that are not part of System.DateTime.

This lcommit also adds a constructor or two.

This commit also adds an ITimeSpan interface.

This commit contains test for TimeSpan that are not implemented.

This commit adds example code for how to use.
This code is commented in wait for internet access.