(Permutation)Add ToTestData method to make it easier to write test data
This commit adds a method that should make it easier to write test data.
Before the test data was written something like:
```
public static IEnumerable<object[]> Variants()
{
    return Permutation.Permutate(
            Permutation.AllIEnumItems<WebPage>(),
            Permutation.AllBools(),
            Permutation.AllBools()
        )
        .Select(data =>
        {
            return new[] {
                data.First(),
                data.Skip(1).First(),
                data.Skip(2).First()
            };
        });
}
```
after this commit it looks like
```
public static IEnumerable<object[]> TestData()
{
    IEnumerable<IEnumerable<object>> res = Permutation.Permutate(
        new object[] { 1, 2 },
        new object[] { "a", "b" }
    );

    return res.ToTestData();
}
```
Okok, the testdata differs in above example. But it is clear that the last
linq Select call is replaced with a simple ToTestData call.

This commit is the first part of #62.
There are some refining of existing test to come.