AssemblyAbstractions:Get rid of build warnings/Update to dotnet6
This commit gets rid of all build warnings to make a cleaner output when compiling.

One of the warnings was about dotnet5 going end-of-life
and hence being updated.
This is a breaking change.

As the nuspec file as gone missing there is no new nuget package.

This commit is part of #72.