DateTImeAbstractions:Get rid of build warnings
This commit gets rid of some build warnings to make a cleaner output when compiling.

There is muck more work to be done in this project
and the whole project might have the wrong design.
So I won't put too much effort into it. Presently.

This commit is part of #72.