Build:Revert to use brandedoutcast/publish-nuget@v2.5.5 again
This commit reverts the nuget publish module back to brandedoutcast
even though it is obsolete
because the newer alirezanet
did not solve the problem
with nuget publishing showing an error
while still working.
You see, even though the build/publish shines red
the package is still published.
Very annoying but it seems to work. For now.