CharacterSeparated:Parse doubles and strings with leading space as respective type with implicitString==true
This commit is a breaking change.

" false" is now parsed as (bool)false when implicitString==true.
" 1.0" is now parsed as (double)1.0 when implicitString==true.

The earlier implementation could not handle " false" and " 1.0" in a "natural" way
but let them always be parsed as string.

There is another caveat; also a breaking change.
We required the decimal character to be ".", as is parsed with CultureInfo.InvariantCulture.
This might change in the future and be cultural aware or, settable, or both.

The testing framework i changed from MSTest to XUnit.
The reason for this is 1) to align all test frameworks in CompulsoryCow
and 2) make ncrunch be able to run the test. One test showed up as red
on the developer machine but green on the build machine.
When the change was made the issue of how to treat decimal character
surfaced.
With more care this could be broken into separate PRs
but there is only so much time in a day.

This commit solves #95.